fikset mappestruktur og navngivning
