+++ src/components/layouts/nve-box/nve-box.styles.ts
