+++ src/components/layouts/nve-stack/nve-stack.styles.ts
