+++ src/nve-designsystem.ts
@@ -4,10 +4,10 @@ export { default as NveAccordion } from './components/nve-accordion/nve-accordio
 export { default as NveAccordionItem } from './components/nve-accordion-item/nve-accordion-item.component';
 export { default as NveAlert } from './components/nve-alert/nve-alert.component';
 export { default as NveBadge } from './components/nve-badge/nve-badge.component';
-export { default as BoxLayout } from './components/box-layout/box-layout.component';
-export { default as ClusterLayout } from './components/cluster-layout/cluster-layout.component';
-export { default as GridLayout } from './components/grid-layout/grid-layout.component';
-export { default as StackLayout } from './components/stack-layout/stack-layout.component';
+export { default as NveBox } from './components/layouts/nve-box/nve-box.component';
+export { default as NveCluster } from './components/layouts/nve-cluster/nve-cluster.component';
+export { default as NveGrid } from './components/layouts/nve-grid/nve-grid.component';
+export { default as NveStack } from './components/layouts/nve-stack/nve-stack.component';
 export { default as NveButton } from './components/nve-button/nve-button.component';
 export { default as NveCarousel } from './components/nve-carousel/nve-carousel.component';
 export { default as NveCarouselItem } from './components/nve-carousel-item/nve-carousel-item.component';
